feat: Separaçao dos CRUDs em arquivos diferentes
