+++ src/menu/sistema/controle/crudmanagers/RespostasCRUD.java
@@ -1,0 +1,137 @@
+package menu.sistema.controle.crudmanagers;
+
+import crud.Crud;
+import menu.pergunta.indices.ListaIDs;
+import produtos.Resposta;
+
+public class RespostasCRUD {
+
+    // Path dos Cruds
+	private final String		    path;
+
+    private static Crud<Resposta>   respostas;
+    private static ListaIDs         respostasUsuario;
+    private static ListaIDs 		perguntasRespostas;
+
+    public RespostasCRUD(String path) {
+
+		this.path = path;
+        try {
+
+            respostas          = new Crud<>("Respostas",  Resposta.class.getConstructor());
+			respostasUsuario   = new ListaIDs(this.path + "/" + "respostasUsuarioIDs");
+            perguntasRespostas = new ListaIDs(this.path + "/" + "perguntasRespostasIDs");
+        } catch(Exception e) {e.printStackTrace();}
+    }
+    
+	/**
+	 * Armazena uma nova resposta nos bancos de dados
+	 * @param r é a resposta a ser armazenada
+	 * @param idPergunta é a pergunta na qual a resposta corresponde
+	 * @return
+	 */
+	public int inserir(Resposta r, int idPergunta) {
+
+		int idResposta = -1;
+
+		idResposta = respostas.create(r);
+		r.setId(idResposta);
+
+		respostasUsuario.create(r.getIdUsuario(),r.getId());
+		perguntasRespostas.create(idPergunta,r.getId());
+
+		return idResposta;
+	}
+
+	/**
+	 * Função para achar uma resposta no banco de dados a partir da ID
+	 * @param id que é a ID da resposta
+	 * @return a resposta correspondente a ID (caso ela exista)
+	 */
+	public static Resposta achar(int id) {
+
+		Resposta r = null;
+		try {
+			r = respostas.read(id);
+		} catch(Exception e ) { e.printStackTrace(); }
+		return r;
+
+	}
+
+	/**
+	 * Função para atualizar uma resposta
+	 * @param r é a pergunta a ser atualizada
+	 */
+	public void atualizar(Resposta r) { 
+
+		respostas.update(r,r.getId());
+	}
+
+	/**
+	 * Função para pegar um array de respostas com base na Id de um usuário no banco de dados
+	 * @param idPergunta da pergunta
+	 * @param idUsuario do usuário que requisitou o processo
+	 * @return array de respostas do idUsuario(se ele tiver registrado)
+	 */
+	public static Resposta[] getRespostaArrayUser(int idUsuario) {
+
+		Resposta[]   resp  		     = null;
+		int[] 	     idsRespostas    = null;
+		
+		idsRespostas = respostasUsuario.read(idUsuario);
+
+        if(idsRespostas == null)
+			return  null;
+
+		resp = new Resposta[idsRespostas.length];
+
+		int contador = 0;
+        for (int i : idsRespostas) {
+            try {
+                Resposta temp = respostas.read(i);
+                if(temp == null)
+					continue;
+
+				resp[contador] = temp; 
+				contador++;
+            }catch(Exception e) { e.printStackTrace(); }
+		}
+
+		return resp;
+	}
+
+	/**
+	 * Função para pegar um array de respostas com base na Id de uma pergunta no banco de dados
+	 * @param idPergunta da pergunta
+	 * @return array de respostas do idUsuario(se ele tiver registrado)
+	 */
+	public static Resposta[] getRespostaArrayGeral(int idPergunta) {
+
+		Resposta[]   resp  		     = null;
+		int[] 	     idsRespostas    = null;
+		
+		idsRespostas = perguntasRespostas.read(idPergunta);
+
+        if(idsRespostas == null)
+			return  null;
+
+		resp = new Resposta[idsRespostas.length];
+
+		int contador = 0;
+        for (int i : idsRespostas) {
+            try {
+                Resposta temp = respostas.read(i);
+                if(temp == null)
+					continue;
+				if(temp.getIdPergunta() != idPergunta) {
+					System.out.println("ue");
+				}
+
+				resp[contador] = temp; 
+				contador++;
+            } catch(Exception e) { e.printStackTrace(); }
+		}
+
+		return resp;
+	}
+}
